+++ packages/backend/src/config/parseConfig.ts
@@ -6,6 +6,7 @@ import { ParseError } from '../errors';
 export enum ProjectType {
     DBT = 'dbt',
     DBT_REMOTE_SERVER = 'dbt_remote_server',
+    DBT_CLOUD_IDE = 'dbt_cloud_ide',
 }
 
 interface DbtProjectConfigBase {
@@ -27,11 +28,25 @@ interface DbtRemoteProjectConfig extends DbtProjectConfigBase {
     rpc_server_port: number;
 }
 
-export type DbtProjectConfig = DbtLocalProjectConfig | DbtRemoteProjectConfig;
+interface DbtCloudIDEProjectConfig extends DbtProjectConfigBase {
+    type: ProjectType.DBT_CLOUD_IDE;
+    api_key: string;
+    account_id: string | number;
+    environment_id: string | number;
+    project_id: string | number;
+}
+
+export type DbtProjectConfig =
+    | DbtLocalProjectConfig
+    | DbtRemoteProjectConfig
+    | DbtCloudIDEProjectConfig;
+
+export type DbtProjectConfigIn<T extends DbtProjectConfig> = Partial<T> &
+    DbtProjectConfigBase;
 
 export type LightdashConfigIn = {
     version: '1.0';
-    projects: Array<DbtProjectConfigBase>;
+    projects: Array<Partial<DbtProjectConfig> & DbtProjectConfigBase>;
 };
 
 export type LightdashConfig = {
@@ -39,24 +54,18 @@ export type LightdashConfig = {
     projects: Array<DbtProjectConfig>;
 };
 
-const isDbtLocalProjectConfig = (
-    config: DbtProjectConfigBase,
-): config is DbtLocalProjectConfig => config.type === ProjectType.DBT;
-
-const isDbtRemoteProjectConfig = (
-    config: DbtProjectConfigBase,
-): config is DbtRemoteProjectConfig =>
-    config.type === ProjectType.DBT_REMOTE_SERVER;
-
 const dbtLocalProjectConfigRequiredFields: Array<keyof DbtLocalProjectConfig> =
-    ['type', 'name', 'profiles_dir', 'project_dir', 'rpc_server_port'];
+    ['profiles_dir', 'project_dir', 'rpc_server_port'];
 const dbtRemoteProjectConfigRequiredFields: Array<
     keyof DbtRemoteProjectConfig
-> = ['type', 'name', 'rpc_server_host', 'rpc_server_port'];
+> = ['rpc_server_host', 'rpc_server_port'];
+const dbtCloudIdeProjectConfigRequiredFields: Array<
+    keyof DbtCloudIDEProjectConfig
+> = ['api_key', 'account_id', 'environment_id', 'project_id'];
 
 const mergeProjectWithEnvironment = <T extends DbtProjectConfig>(
     projectIndex: number,
-    project: T,
+    project: DbtProjectConfigIn<T>,
     requiredField: Array<keyof T>,
 ): T =>
     requiredField.reduce((prev, key) => {
@@ -72,28 +81,38 @@ const mergeProjectWithEnvironment = <T extends DbtProjectConfig>(
             ...prev,
             [key]: value,
         };
-    }, project);
+    }, project) as T;
 
 const mergeWithEnvironment = (config: LightdashConfigIn): LightdashConfig => {
     const mergedProjects = config.projects.map((project, idx) => {
-        if (isDbtLocalProjectConfig(project)) {
-            return mergeProjectWithEnvironment(
-                idx,
-                project,
-                dbtLocalProjectConfigRequiredFields,
-            );
+        const projectType = project.type;
+        switch (project.type) {
+            case ProjectType.DBT:
+                return mergeProjectWithEnvironment(
+                    idx,
+                    project,
+                    dbtLocalProjectConfigRequiredFields,
+                );
+            case ProjectType.DBT_REMOTE_SERVER:
+                return mergeProjectWithEnvironment(
+                    idx,
+                    project,
+                    dbtRemoteProjectConfigRequiredFields,
+                );
+            case ProjectType.DBT_CLOUD_IDE:
+                return mergeProjectWithEnvironment(
+                    idx,
+                    project,
+                    dbtCloudIdeProjectConfigRequiredFields,
+                );
+            default: {
+                const never: never = project;
+                throw new ParseError(
+                    `Lightdash config file successfully loaded but invalid: Project type: ${projectType} is not supported`,
+                    {},
+                );
+            }
         }
-        if (isDbtRemoteProjectConfig(project)) {
-            return mergeProjectWithEnvironment(
-                idx,
-                project,
-                dbtRemoteProjectConfigRequiredFields,
-            );
-        }
-        throw new ParseError(
-            `Lightdash config file successfully loaded but invalid: Project type: ${project.type} is not supported`,
-            {},
-        );
     });
     return {
         ...config,
@@ -106,6 +125,7 @@ export const parseConfig = (raw: any): LightdashConfig => {
         schemaId: 'id',
         useDefaults: true,
         discriminator: true,
+        allowUnionTypes: true,
     });
     addFormats(ajv);
     const validate = ajv.compile<LightdashConfigIn>(lightdashV1JsonSchema);