+++ packages/backend/src/projectAdapters/dbtRemoteProjectAdapter.ts
@@ -1,9 +1,15 @@
 import { Explore } from 'common';
 import { DbtBaseProjectAdapter } from './dbtBaseProjectAdapter';
+import { DbtRpcClient } from '../dbt/dbtRpcClient';
 
 export class DbtRemoteProjectAdapter extends DbtBaseProjectAdapter {
-    constructor(host: string, port: number) {
-        super(`http://${host}:${port}/jsonrpc`);
+    rpcClient: DbtRpcClient;
+
+    constructor(host: string, port: number, protocol: string = 'http') {
+        super();
+        this.rpcClient = new DbtRpcClient(
+            `${protocol}://${host}:${port}/jsonrpc`,
+        );
     }
 
     public async compileAllExplores(): Promise<Explore[]> {